refactor: extract domain logic from application layer
- Move event creation logic to domain layer
  - addCoursesToSession now returns CoursesAddedToSession event
  - createRegistrationSession function for event creation
- Simplify application command handlers
  - Remove validation builder in favor of sequential execution
  - Commands now only handle infrastructure concerns
- Add comprehensive JSDoc documentation
- Delete unused validation-builder.ts

This refactoring improves separation of concerns by ensuring domain
logic remains in the domain layer while application layer focuses on
orchestration and infrastructure coordination.
