+++ src/contexts/enrollment/application/commands/create-registration-session.ts
@@ -1,7 +1,7 @@
 import { Effect } from "effect";
 import { StudentId, Term, RegistrationSessionId } from "../../domain/models/shared/value-objects.js";
 import { SessionAlreadyExists } from "../../domain/errors/domain-errors.js";
-import { RegistrationSessionCreated } from "../../domain/events/registration-session-events.js";
+import { createRegistrationSession as createRegistrationSessionEvent } from "../../domain/models/registration-session/registration-session.js";
 import { RegistrationSessionRepository } from "../../domain/repositories/registration-session-repository.js";
 import { EventStore } from "../../../shared/kernel/types/event-store.js";
 import { EventBus } from "../../../shared/kernel/types/event-bus.js";
@@ -11,7 +11,11 @@ export interface CreateRegistrationSessionCommand {
   readonly term: Term;
 }
 
-// セッションが既に存在しないことを確認するヘルパー関数
+/**
+ * セッションが存在しないことを確認するヘルパー関数
+ * @param sessionId - 確認するセッションのID
+ * @returns - セッションが存在しない場合は何も返さず、存在する場合はSessionAlreadyExistsエラーを投げる
+ */
 const ensureNotExists = (sessionId: RegistrationSessionId) =>
   Effect.gen(function* () {
     const repository = yield* RegistrationSessionRepository;
@@ -25,6 +29,12 @@ const ensureNotExists = (sessionId: RegistrationSessionId) =>
     );
   });
 
+/**
+ * 履修登録セッションを作成するコマンド
+ * @param command - 作成するセッションの情報
+ * @returns - 作成されたセッションのID
+ * @remark この関数は、セッションの状態を変更するのではなく、ドメインイベントを生成して保存・パブリッシュします。
+ */
 export const createRegistrationSession = (
   command: CreateRegistrationSessionCommand
 ) =>
@@ -37,19 +47,11 @@ export const createRegistrationSession = (
     // TODO: studentIdで学生の存在を確認するロジックを追加する
     // TODO: termの妥当性を確認するロジックを追加する
 
-    // 複合キーからセッションIDを生成
     const sessionId = yield* RegistrationSessionId.create(studentId, term);
 
-    // 既存セッションが存在しないことを確認（複合キーで直接チェック）
     yield* ensureNotExists(sessionId);
 
-    // ドメインイベントを作成
-    const event = new RegistrationSessionCreated({
-      sessionId,
-      studentId,
-      term,
-      createdAt: new Date()
-    });
+    const event = createRegistrationSessionEvent(sessionId, studentId, term);
 
     // イベントを保存（イベントソーシング: セッションは永続化せずイベントのみ保存）
     // 注意: 現在は非同期投影を実装していないため、リポジトリはイベントから再構築
